Newsagent/requirements.txt
# News Agent Requirements
# Python dependencies for the Autonomous AI News Agent

# Core dependencies
python-dotenv==1.0.0
requests==2.31.0
feedparser==6.0.10

# LangChain and AI
langchain==0.3.7
langchain-openai==0.2.5
langchain-core==0.3.15
openai==1.54.0

# Data processing
python-dateutil==2.8.2

# Streamlit UI
streamlit==1.29.0
pandas==2.1.4

# Optional but recommended
beautifulsoup4==4.12.2
lxml==4.9.3
newspaper3k==0.2.8

# Development tools (optional)
pytest==7.4.3
black==23.12.1
flake8==6.1.0
